backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
faiss-cpu
python-dotenv
langchain-community
sentence-transformers
pydantic
